Remove export_metadata module
